Выполнил задание 4
