+++ src/ObjectOrientedPractics/src/Model/PriorityOrder.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Text.RegularExpressions;
 using static ObjectOrientedPractics.Model.Store;
 
 namespace ObjectOrientedPractics.Model
@@ -8,6 +10,16 @@ namespace ObjectOrientedPractics.Model
     /// </summary>
     public class PriorityOrder : Order
     {
+        public static List<DeliveryTimeInterval> DeliveryTimeIntervals { get; } =
+            new List<DeliveryTimeInterval>
+            {
+                new DeliveryTimeInterval("09:00 - 11:00"),
+                new DeliveryTimeInterval("11:00 - 13:00"),
+                new DeliveryTimeInterval("13:00 - 15:00"),
+                new DeliveryTimeInterval("15:00 - 17:00"),
+                new DeliveryTimeInterval("17:00 - 19:00")
+            };
+
         public PriorityOrder()
         {
             DeliveryDate = DateTime.Now;
@@ -23,5 +35,33 @@ namespace ObjectOrientedPractics.Model
         /// Время доставки.
         /// </summary>
         public DeliveryTimeInterval DeliveryInterval { get; set; }
+
+        public class DeliveryTimeInterval
+        {
+            private static readonly Regex Regex =
+                new Regex(@"\d{2}:\d{2} - \d{2}:\d{2}");
+
+            public string Value { set; get; }
+
+            public DeliveryTimeInterval(string value)
+            {
+                Value = value;
+            }
+
+            public override string ToString()
+            {
+                return Value;
+            }
+
+            public static DeliveryTimeInterval ValueOf(string value)
+            {
+                if (Regex.IsMatch(value))
+                {
+                    return DeliveryTimeIntervals.Find(interval => interval.Value.Equals(value));
+                }
+
+                throw new ArgumentException("Argument not valid!");
+            }
+        }
     }
 }
